+++ stts/android/src/main/kotlin/com/llfbandit/stts/tts/model/TtsOptions.kt
@@ -1,0 +1,7 @@
+package com.llfbandit.stts.tts.model
+
+data class TtsOptions(
+  val queueMode: TtsQueueMode,
+  val preSilenceMs: Int?,
+  val postSilenceMs: Int?
+)