+++ stts_platform_interface/lib/src/tts/model/tts_options.dart
@@ -2,8 +2,22 @@ import 'tts_queue_mode.dart';
 
 /// Text-to-Speech options
 class TtsOptions {
-  /// Queue behaviour mode (flush or add)
+  /// Queue behaviour mode (flush or add/enqueue).
   final TtsQueueMode mode;
 
-  const TtsOptions({this.mode = TtsQueueMode.add});
+  /// Delay (silence) before speaking the utterance.
+  ///
+  /// *Ignored on web platform.*
+  final Duration? preSilence;
+
+  /// Delay (silence) after speaking the utterance.
+  ///
+  /// *Ignored on web platform.*
+  final Duration? postSilence;
+
+  const TtsOptions({
+    this.mode = TtsQueueMode.add,
+    this.preSilence,
+    this.postSilence,
+  });
 }